Fix Messages.sh error from me from yesterday. Patch by Burkhard Lueck, thanks.

+++ plugins/Messages.sh
@@ -1,5 +1,5 @@
 #! /usr/bin/env bash
-for plugin in `find -maxdepth 1 -type d -not -name . -printf "%f\n"` ; do
+for plugin in `find -maxdepth 1 -type d -not -name . -not -name .svn -printf "%f\n"` ; do
 
     rcfiles="`find $plugin -name \*.rc`"
     uifiles="`find $plugin -name \*.ui`"
@@ -7,8 +7,12 @@ for plugin in `find -maxdepth 1 -type d -not -name . -printf "%f\n"` ; do
         $EXTRACTRC "$rcfiles" >> rc.cpp || exit 11
     fi
     if [[ "$uifiles" != "" ]] ; then
-        $EXTRACTRC  >> rc.cpp || exit 12
+        $EXTRACTRC "$uifiles" >> rc.cpp || exit 12
     fi
-    $XGETTEXT -kaliasLocal `find $plugin -name \*.cc -o -name \*.cpp -o -name \*.h` rc.cpp -o $podir/kdev${plugin}.pot
-    rm -f rc.cpp
+    if [ -f rc.cpp ] ; then
+        $XGETTEXT -kaliasLocal `find $plugin -name \*.cc -o -name \*.cpp -o -name \*.h` rc.cpp -o $podir/kdev${plugin}.pot
+        rm -f rc.cpp
+    else
+        $XGETTEXT -kaliasLocal `find $plugin -name \*.cc -o -name \*.cpp -o -name \*.h` -o $podir/kdev${plugin}.pot
+    fi
 done
